@@ -16,17 +16,18 @@
   "author": "Dian Dimitrov",
   "license": "MIT",
   "dependencies": {
-    "babel-loader": "^6.2.2",
-    "babel-plugin-transform-class-properties": "^6.4.0",
-    "babel-plugin-transform-runtime": "^6.4.3",
+    "babel-loader": "^6.2.5",
+    "babel-plugin-transform-class-properties": "^6.11.5",
+    "babel-plugin-transform-runtime": "^6.12.0",
     "babel-polyfill": "^6.3.14",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
     "babel-register": "^6.4.3",
+    "babel-runtime": "^6.11.6",
     "express": "^4.13.4",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
+    "react": "^15.3.1",
+    "react-dom": "^15.3.1",
     "react-router": "^2.7.0",
     "webpack": "^1.12.13"
   },
